17.12.20

+++ src/org/orgname/app/ui/TableForm.java
@@ -5,13 +5,15 @@ import org.orgname.app.database.entity.ClientEntity;
 import org.orgname.app.database.manager.ClientEntityManager;
 import org.orgname.app.util.BaseForm;
 import org.orgname.app.util.CustomTableModel;
+import org.orgname.app.util.DialogUtil;
 
 import javax.swing.*;
-import java.awt.event.KeyAdapter;
-import java.awt.event.KeyEvent;
-import java.awt.event.MouseAdapter;
-import java.awt.event.MouseEvent;
+import java.awt.event.*;
 import java.sql.SQLException;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.Comparator;
+import java.util.List;
 
 public class TableForm extends BaseForm
 {
@@ -19,12 +21,28 @@ public class TableForm extends BaseForm
 
     private CustomTableModel<ClientEntity> model;
 
+    private boolean idSort = true;
+
     private JPanel mainPanel;
     private JTable table;
+    private JButton idSortButton;
+    private JButton bithdaySortButton;
+    private JComboBox genderSortBox;
 
     public TableForm()
     {
         setContentPane(mainPanel);
+
+        initTable();
+        initButtons();
+        initBoxes();
+
+        setVisible(true);
+    }
+
+    private void initTable()
+    {
+        table.getTableHeader().setReorderingAllowed(false);
 
         try {
             model = new CustomTableModel<>(
@@ -51,8 +69,12 @@ public class TableForm extends BaseForm
                 @Override
                 public void keyPressed(KeyEvent e) {
                     int row = table.getSelectedRow();
-                    if(row != -1)
+                    if(e.getKeyCode() == KeyEvent.VK_DELETE && row != -1)
                     {
+                        if(!DialogUtil.showConfirm(TableForm.this, "ВЫ точно хоиите удалить данную запись?")) {
+                            return;
+                        }
+
                         try {
                             clientEntityManager.delete(model.getValues().get(row));
                             model.getValues().remove(row);
@@ -68,8 +90,73 @@ public class TableForm extends BaseForm
         } catch (SQLException throwables) {
             throwables.printStackTrace();
         }
+    }
 
-        setVisible(true);
+    private void initButtons()
+    {
+        idSortButton.addActionListener(e -> {
+            Collections.sort(model.getValues(), new Comparator<ClientEntity>() {
+                @Override
+                public int compare(ClientEntity o1, ClientEntity o2) {
+                    if(!idSort) {
+                        return Integer.compare(o1.getId(), o2.getId());
+                    } else {
+                        return Integer.compare(o2.getId(), o1.getId());
+                    }
+                }
+            });
+            idSort = !idSort;
+            model.fireTableDataChanged();
+        });
+
+        bithdaySortButton.addActionListener(e -> {
+            Collections.sort(model.getValues(), new Comparator<ClientEntity>() {
+                @Override
+                public int compare(ClientEntity o1, ClientEntity o2) {
+                    return o1.getBirthday().compareTo(o2.getBirthday());
+                }
+            });
+            model.fireTableDataChanged();
+        });
+    }
+
+    private void initBoxes()
+    {
+        genderSortBox.addItem("Все гендеры");
+        genderSortBox.addItem("Мужской");
+        genderSortBox.addItem("Женский");
+
+        genderSortBox.addItemListener(e -> {
+            if(e.getStateChange() == ItemEvent.SELECTED) {
+
+                try {
+                    List<ClientEntity> allClients = clientEntityManager.getAll();
+                    int selected = genderSortBox.getSelectedIndex();
+                    if(selected == 0) {
+                        model.setValues(allClients);
+                    } else if(selected == 1) {
+                        List<ClientEntity> temp = new ArrayList<>();
+                        for(ClientEntity c : allClients) {
+                            if(c.getGenderCode() == 'м') {
+                                temp.add(c);
+                            }
+                        }
+                        model.setValues(temp);
+                    } else if(selected == 2) {
+                        List<ClientEntity> temp = new ArrayList<>();
+                        for(ClientEntity c : allClients) {
+                            if(c.getGenderCode() == 'ж') {
+                                temp.add(c);
+                            }
+                        }
+                        model.setValues(temp);
+                    }
+                    model.fireTableDataChanged();
+                } catch (SQLException throwables) {
+                    throwables.printStackTrace();
+                }
+            }
+        });
     }
 
     @Override
@@ -80,5 +167,9 @@ public class TableForm extends BaseForm
     @Override
     public int getFormHeight() {
         return 600;
+    }
+
+    public CustomTableModel<ClientEntity> getModel() {
+        return model;
     }
 }
